# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for GoPOS Backend",
        "title": "GoPOS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/shifts/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "End the current active shift, calculate total expected cash, and record physical cash discrepancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Close an active cashier shift",
                "parameters": [
                    {
                        "description": "Closing Shift Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CloseShiftInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift successfully closed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "No active shift found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/shifts/open": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new shift for the authenticated cashier by recording the initial drawer cash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Open a new cashier shift",
                "parameters": [
                    {
                        "description": "Initial Shift Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.OpenShiftInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shift successfully opened",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Active shift already exists or invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/transactions": {
            "get": {
                "description": "Fetch a historical data view of invoice logs with standard offset pagination parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page target number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max capacity per page response window",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully returned transaction lists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Database mapping reading error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "description": "Fetch all purchase rows and overall financial totals of a transaction by its numeric unique key identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get invoice detail sheet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Numeric Internal Transaction Database ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target database row matched and loaded",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invoice not found matching specific identifier",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CloseShiftInput": {
            "type": "object",
            "required": [
                "total_cash_actual"
            ],
            "properties": {
                "total_cash_actual": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "handlers.OpenShiftInput": {
            "type": "object",
            "required": [
                "start_cash"
            ],
            "properties": {
                "start_cash": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}